django1/requirements.txt
Django>=4.0
djangorestframework
drf-spectacular